Dual complexes

API docs

CombinatorialSpaces.DualSimplicialSets — Module

Dual complexes for simplicial sets in one, two, and three dimensions.

CombinatorialSpaces.DualSimplicialSets.AbstractDeltaDualComplex1D — Type

Abstract type for dual complex of a 1D delta set.

CombinatorialSpaces.DualSimplicialSets.AbstractDeltaDualComplex2D — Type

Abstract type for dual complex of a 2D delta set.

CombinatorialSpaces.DualSimplicialSets.Barycenter — Type

Barycenter, aka centroid, of a simplex.

CombinatorialSpaces.DualSimplicialSets.Circumcenter — Type

Circumcenter, or center of circumscribed circle, of a simplex.

The circumcenter is calculated by inverting the Cayley-Menger matrix, as explained by Westdendorp. This method of calculation is also used in the package AlphaShapes.jl.

CombinatorialSpaces.DualSimplicialSets.DeltaDualComplex1D — Type

Dual complex of a one-dimensional delta set.

The data structure includes both the primal complex and the dual complex, as well as the mapping between them.

CombinatorialSpaces.DualSimplicialSets.DeltaDualComplex2D — Type

Dual complex of a two-dimensional delta set.

CombinatorialSpaces.DualSimplicialSets.DualChain — Type

Wrapper for chain of dual cells of dimension n.

In an $N$-dimensional complex, the elementary dual simplices of each $n$-simplex together comprise the dual $(N-n)$-cell of the simplex. Using this correspondence, a basis for primal $n$-chains defines the basis for dual $(N-n)$-chains.

Note

In (Hirani 2003, Definition 3.4.1), the duality operator assigns a certain sign to each elementary dual simplex. For us, all of these signs should be regarded as positive because we have already incorporated them into the orientation of the dual simplices.

CombinatorialSpaces.DualSimplicialSets.DualE — Type

Edge in simplicial set: alias for Simplex{1}.

CombinatorialSpaces.DualSimplicialSets.DualForm — Type

Wrapper for form, aka cochain, on dual cells of dimension n.

CombinatorialSpaces.DualSimplicialSets.DualSimplex — Type

Wrapper for dual simplex or simplices of dimension D.

See also: DualV, DualE, DualTri.

CombinatorialSpaces.DualSimplicialSets.DualTri — Type

Triangle in simplicial set: alias for Simplex{2}.

CombinatorialSpaces.DualSimplicialSets.DualV — Type

Vertex in simplicial set: alias for Simplex{0}.

CombinatorialSpaces.DualSimplicialSets.DualVectorField — Type

Wrapper for vector field on dual vertices.

CombinatorialSpaces.DualSimplicialSets.EmbeddedDeltaDualComplex1D — Type

Embedded dual complex of an embedded 1D delta set.

Although they are redundant information, the lengths of the primal and dual edges are precomputed and stored.

CombinatorialSpaces.DualSimplicialSets.EmbeddedDeltaDualComplex2D — Type

Embedded dual complex of an embedded 12 delta set.

Although they are redundant information, the lengths and areas of the primal/dual edges and triangles are precomputed and stored.

CombinatorialSpaces.DualSimplicialSets.Incenter — Type

Incenter, or center of inscribed circle, of a simplex.

CombinatorialSpaces.DualSimplicialSets.OrientedDeltaDualComplex1D — Type

Oriented dual complex of an oriented 1D delta set.

CombinatorialSpaces.DualSimplicialSets.OrientedDeltaDualComplex2D — Type

Oriented dual complex of an oriented 2D delta set.

CombinatorialSpaces.DualSimplicialSets.SimplexCenter — Type

A notion of "geometric center" of a simplex.

See also: geometric_center.

CombinatorialSpaces.DualSimplicialSets.:∧ — Method

Wedge product of discrete forms.

The wedge product of a $k$-form and an $l$-form is a $(k+l)$-form.

The DEC and related systems have several flavors of wedge product. This one is the discrete primal-primal wedge product introduced in (Hirani, 2003, Chapter 7) and (Desbrun et al 2005, Section 8). It depends on the geometric embedding and requires the dual complex.

CombinatorialSpaces.DualSimplicialSets.:⋆ — Method

Hodge star operator from primal $n$-forms to dual $N-n$-forms.

Warning

Some authors, such as (Hirani 2003) and (Desbrun 2005), use the symbol $⋆$ for the duality operator on chains and the symbol $*$ for the Hodge star operator on cochains. We do not explicitly define the duality operator and we use the symbol $⋆$ for the Hodge star.

CombinatorialSpaces.DualSimplicialSets.codifferential — Function

Alias for the codifferential operator δ.

CombinatorialSpaces.DualSimplicialSets.dual_boundary — Method

Boundary of chain of dual cells.

Transpose of dual_derivative.

CombinatorialSpaces.DualSimplicialSets.dual_derivative — Method

Discrete exterior derivative of dual form.

Transpose of dual_boundary. For more info, see (Desbrun, Kanso, Tong, 2008: Discrete differential forms for computational modeling, 4.5).

CombinatorialSpaces.DualSimplicialSets.dual_point — Method

Point associated with dual vertex of complex.

CombinatorialSpaces.DualSimplicialSets.dual_triangle_vertices — Method

Boundary dual vertices of a dual triangle.

This accessor assumes that the simplicial identities for the dual hold.

CombinatorialSpaces.DualSimplicialSets.edge_center — Method

Dual vertex corresponding to center of primal edge.

CombinatorialSpaces.DualSimplicialSets.elementary_duals — Method

List of elementary dual simplices corresponding to primal simplex.

In general, in an $n$-dimensional complex, the elementary duals of primal $k$-simplices are dual $(n-k)$-simplices. Thus, in 1D dual complexes, the elementary duals of...

primal vertices are dual edges

primal edges are (single) dual vertices

In 2D dual complexes, the elementary duals of...

primal vertices are dual triangles

primal edges are dual edges

primal triangles are (single) dual triangles

CombinatorialSpaces.DualSimplicialSets.flat — Function

Alias for the flat operator ♭.

CombinatorialSpaces.DualSimplicialSets.geometric_center — Function

Calculate the center of simplex spanned by given points.

The first argument is a list of points and the second specifies the notion of "center", via an instance of SimplexCenter.

CombinatorialSpaces.DualSimplicialSets.hodge_star — Function

Alias for the Hodge star operator ⋆.

CombinatorialSpaces.DualSimplicialSets.interior_product — Method

Interior product of a vector field (or 1-form) and a $n$-form.

Specifically, this operation is the primal-dual interior product defined in (Hirani 2003, Section 8.2) and (Desbrun et al 2005, Section 10). Thus it takes a primal vector field (or primal 1-form) and a dual $n$-forms and then returns a dual $(n-1)$-form.

CombinatorialSpaces.DualSimplicialSets.interior_product_flat — Method

Interior product of a 1-form and a $n$-form, yielding an $(n-1)$-form.

Usually, the interior product is defined for vector fields; this function assumes that the flat operator ♭ (not yet implemented for primal vector fields) has already been applied to yield a 1-form.

CombinatorialSpaces.DualSimplicialSets.laplace_beltrami — Function

Alias for the Laplace-Beltrami operator Δ.

CombinatorialSpaces.DualSimplicialSets.lie_derivative — Method

Lie derivative of $n$-form with respect to a vector field (or 1-form).

Specifically, this is the primal-dual Lie derivative defined in (Hirani 2003, Section 8.4) and (Desbrun et al 2005, Section 10).

CombinatorialSpaces.DualSimplicialSets.lie_derivative_flat — Method

Lie derivative of $n$-form with respect to a 1-form.

Assumes that the flat operator ♭ has already been applied to the vector field.

CombinatorialSpaces.DualSimplicialSets.primal_vertex — Method

Primal vertex associated with a dual simplex.

CombinatorialSpaces.DualSimplicialSets.subdivide_duals! — Method

Compute geometric subdivision for embedded dual complex.

Supports different methods of subdivision through the choice of geometric center, as defined by geometric_center. In particular, barycentric subdivision and circumcentric subdivision are supported.

CombinatorialSpaces.DualSimplicialSets.subsimplices — Method

List of dual simplices comprising the subdivision of a primal simplex.

A primal $n$-simplex is always subdivided into $n!$ dual $n$-simplices, not be confused with the elementary_duals which have complementary dimension.

The returned list is ordered such that subsimplices with the same primal vertex appear consecutively.

CombinatorialSpaces.DualSimplicialSets.triangle_center — Method

Dual vertex corresponding to center of primal triangle.

CombinatorialSpaces.DualSimplicialSets.vertex_center — Method

Dual vertex corresponding to center of primal vertex.

CombinatorialSpaces.DualSimplicialSets.wedge_product — Function

Alias for the wedge product operator ∧.

CombinatorialSpaces.DualSimplicialSets.Δ — Method

Laplace-Beltrami operator on discrete forms.

The linear operator on primal $n$-forms defined by $Δ f := δ d f$, where δ is the codifferential and d is the exterior derivative.

CombinatorialSpaces.DualSimplicialSets.δ — Method

Codifferential operator from primal $n$ forms to primal $n-1$-forms.

CombinatorialSpaces.DualSimplicialSets.♭ — Method

Flat operator converting vector fields to 1-forms.

A generic function for the discrete flat operators proposed by (Hirani 2003). Currently only the DPP-flat is implemented.