Discrete exterior calculus

API docs

CombinatorialSpaces.DiscreteExteriorCalculus — Module

The discrete exterior calculus (DEC) for simplicial sets.

This module provides the dual complex associated with a delta set (the primal complex), which is a discrete incarnation of Hodge duality, as well as the many operators of the DEC that depend on it, such as the Hodge star, codifferential, wedge product, interior product, and Lie derivative. The main reference for this module is Hirani's 2003 PhD thesis.

CombinatorialSpaces.DiscreteExteriorCalculus.AbstractDeltaDualComplex1D — Type

Abstract type for dual complex of a 1D delta set.

CombinatorialSpaces.DiscreteExteriorCalculus.AbstractDeltaDualComplex2D — Type

Abstract type for dual complex of a 2D delta set.

CombinatorialSpaces.DiscreteExteriorCalculus.Barycenter — Type

Barycenter, aka centroid, of a simplex.

CombinatorialSpaces.DiscreteExteriorCalculus.Circumcenter — Type

Circumcenter, or center of circumscribed circle, of a simplex.

The circumcenter is calculated by inverting the Cayley-Menger matrix, as explained by Westdendorp. This method of calculation is also used in the package AlphaShapes.jl.

CombinatorialSpaces.DiscreteExteriorCalculus.DeltaDualComplex1D — Type

Dual complex of a one-dimensional delta set.

The data structure includes both the primal complex and the dual complex, as well as the mapping between them.

CombinatorialSpaces.DiscreteExteriorCalculus.DeltaDualComplex2D — Type

Dual complex of a two-dimensional delta set.

CombinatorialSpaces.DiscreteExteriorCalculus.DualChain — Type

Wrapper for chain of dual cells of dimension n.

In an $N$-dimensional complex, the elementary dual simplices of each $n$-simplex together comprise the dual $(N-n)$-cell of the simplex. Using this correspondence, a basis for primal $n$-chains defines the basis for dual $(N-n)$-chains.

Note

In (Hirani 2003, Definition 3.4.1), the duality operator assigns a certain sign to each elementary dual simplex. For us, all of these signs should be regarded as positive because we have already incorporated them into the orientation of the dual simplices.

CombinatorialSpaces.DiscreteExteriorCalculus.DualE — Type

Edge in simplicial set: alias for Simplex{1}.

CombinatorialSpaces.DiscreteExteriorCalculus.DualForm — Type

Wrapper for form, aka cochain, on dual cells of dimension n.

CombinatorialSpaces.DiscreteExteriorCalculus.DualSimplex — Type

Wrapper for dual simplex or simplices of dimension D.

See also: DualV, DualE, DualTri.

CombinatorialSpaces.DiscreteExteriorCalculus.DualTri — Type

Triangle in simplicial set: alias for Simplex{2}.

CombinatorialSpaces.DiscreteExteriorCalculus.DualV — Type

Vertex in simplicial set: alias for Simplex{0}.

CombinatorialSpaces.DiscreteExteriorCalculus.DualVectorField — Type

Wrapper for vector field on dual vertices.

CombinatorialSpaces.DiscreteExteriorCalculus.EmbeddedDeltaDualComplex1D — Type

Embedded dual complex of an embedded 1D delta set.

Although they are redundant information, the lengths of the primal and dual edges are precomputed and stored.

CombinatorialSpaces.DiscreteExteriorCalculus.EmbeddedDeltaDualComplex2D — Type

Embedded dual complex of an embedded 12 delta set.

Although they are redundant information, the lengths and areas of the primal/dual edges and triangles are precomputed and stored.

CombinatorialSpaces.DiscreteExteriorCalculus.Incenter — Type

Incenter, or center of inscribed circle, of a simplex.

CombinatorialSpaces.DiscreteExteriorCalculus.OrientedDeltaDualComplex1D — Type

Oriented dual complex of an oriented 1D delta set.

CombinatorialSpaces.DiscreteExteriorCalculus.OrientedDeltaDualComplex2D — Type

Oriented dual complex of an oriented 2D delta set.

CombinatorialSpaces.DiscreteExteriorCalculus.PrimalVectorField — Type

Wrapper for vector field on primal vertices.

CombinatorialSpaces.DiscreteExteriorCalculus.SimplexCenter — Type

A notion of "geometric center" of a simplex.

See also: geometric_center.

Catlab.Theories.Δ — Method

Laplace-de Rham operator on discrete forms.

This linear operator on primal $n$-forms is defined by $Δ := δ d + d δ$. Restricted to 0-forms, it reduces to the negative of the Laplace-Beltrami operator ∇²: $Δ f = -∇² f$.

CombinatorialSpaces.DiscreteExteriorCalculus.:∧ — Method

Wedge product of discrete forms.

The wedge product of a $k$-form and an $l$-form is a $(k+l)$-form.

The DEC and related systems have several flavors of wedge product. This one is the discrete primal-primal wedge product introduced in (Hirani, 2003, Chapter 7) and (Desbrun et al 2005, Section 8). It depends on the geometric embedding and requires the dual complex.

CombinatorialSpaces.DiscreteExteriorCalculus.:⋆ — Method

Hodge star operator from primal 1-forms to dual 1-forms.

This specific hodge star implementation is based on the hodge star presented in (Ayoub et al 2020), which generalizes the operator presented in (Hirani 2003). This reproduces the diagonal hodge for a dual mesh generated under circumcentric subdivision and provides off-diagonal correction factors for meshes generated under other subdivision schemes (e.g. barycentric).

CombinatorialSpaces.DiscreteExteriorCalculus.:⋆ — Method

Hodge star operator from primal $n$-forms to dual $N-n$-forms.

Note

Some authors, such as (Hirani 2003) and (Desbrun 2005), use the symbol $⋆$ for the duality operator on chains and the symbol $*$ for the Hodge star operator on cochains. We do not explicitly define the duality operator and we use the symbol $⋆$ for the Hodge star.

CombinatorialSpaces.DiscreteExteriorCalculus.codifferential — Function

Alias for the codifferential operator δ.

CombinatorialSpaces.DiscreteExteriorCalculus.dual_boundary — Method

Boundary of chain of dual cells.

Transpose of dual_derivative.

CombinatorialSpaces.DiscreteExteriorCalculus.dual_derivative — Method

Discrete exterior derivative of dual form.

Transpose of dual_boundary. For more info, see (Desbrun, Kanso, Tong, 2008: Discrete differential forms for computational modeling, 4.5).

CombinatorialSpaces.DiscreteExteriorCalculus.dual_point — Method

Point associated with dual vertex of complex.

CombinatorialSpaces.DiscreteExteriorCalculus.dual_triangle_vertices — Method

Boundary dual vertices of a dual triangle.

This accessor assumes that the simplicial identities for the dual hold.

CombinatorialSpaces.DiscreteExteriorCalculus.edge_center — Method

Dual vertex corresponding to center of primal edge.

CombinatorialSpaces.DiscreteExteriorCalculus.elementary_duals — Method

List of elementary dual simplices corresponding to primal simplex.

In general, in an $n$-dimensional complex, the elementary duals of primal $k$-simplices are dual $(n-k)$-simplices. Thus, in 1D dual complexes, the elementary duals of...

primal vertices are dual edges

primal edges are (single) dual vertices

In 2D dual complexes, the elementary duals of...

primal vertices are dual triangles

primal edges are dual edges

primal triangles are (single) dual triangles

CombinatorialSpaces.DiscreteExteriorCalculus.flat — Function

Alias for the flat operator ♭.

CombinatorialSpaces.DiscreteExteriorCalculus.geometric_center — Function

Calculate the center of simplex spanned by given points.

The first argument is a list of points and the second specifies the notion of "center", via an instance of SimplexCenter.

CombinatorialSpaces.DiscreteExteriorCalculus.hodge_star — Function

Alias for the Hodge star operator ⋆.

CombinatorialSpaces.DiscreteExteriorCalculus.interior_product — Method

Interior product of a vector field (or 1-form) and a $n$-form.

Specifically, this operation is the primal-dual interior product defined in (Hirani 2003, Section 8.2) and (Desbrun et al 2005, Section 10). Thus it takes a primal vector field (or primal 1-form) and a dual $n$-forms and then returns a dual $(n-1)$-form.

CombinatorialSpaces.DiscreteExteriorCalculus.interior_product_flat — Method

Interior product of a 1-form and a $n$-form, yielding an $(n-1)$-form.

Usually, the interior product is defined for vector fields; this function assumes that the flat operator ♭ (not yet implemented for primal vector fields) has already been applied to yield a 1-form.

CombinatorialSpaces.DiscreteExteriorCalculus.inv_hodge_star — Method

Inverse Hodge star operator from dual $N-n$-forms to primal $n$-forms.

Confusingly, this is not the operator inverse of the Hodge star ⋆ because it carries an extra global sign, in analogy to the smooth case (Gillette, 2009, Notes on the DEC, Definition 2.27).

CombinatorialSpaces.DiscreteExteriorCalculus.laplace_beltrami — Function

Alias for the Laplace-Beltrami operator ∇².

CombinatorialSpaces.DiscreteExteriorCalculus.laplace_de_rham — Function

Alias for the Laplace-de Rham operator Δ.

CombinatorialSpaces.DiscreteExteriorCalculus.lie_derivative — Function

Alias for Lie derivative operator ℒ.

CombinatorialSpaces.DiscreteExteriorCalculus.lie_derivative_flat — Method

Lie derivative of $n$-form with respect to a 1-form.

Assumes that the flat operator ♭ has already been applied to the vector field.

CombinatorialSpaces.DiscreteExteriorCalculus.primal_vertex — Method

Primal vertex associated with a dual simplex.

CombinatorialSpaces.DiscreteExteriorCalculus.sharp — Function

Alias for the sharp operator ♯.

CombinatorialSpaces.DiscreteExteriorCalculus.subdivide_duals! — Method

Compute geometric subdivision for embedded dual complex.

Supports different methods of subdivision through the choice of geometric center, as defined by geometric_center. In particular, barycentric subdivision and circumcentric subdivision are supported.

CombinatorialSpaces.DiscreteExteriorCalculus.subsimplices — Method

List of dual simplices comprising the subdivision of a primal simplex.

A primal $n$-simplex is always subdivided into $n!$ dual $n$-simplices, not be confused with the elementary_duals which have complementary dimension.

The returned list is ordered such that subsimplices with the same primal vertex appear consecutively.

CombinatorialSpaces.DiscreteExteriorCalculus.triangle_center — Method

Dual vertex corresponding to center of primal triangle.

CombinatorialSpaces.DiscreteExteriorCalculus.vertex_center — Method

Dual vertex corresponding to center of primal vertex.

CombinatorialSpaces.DiscreteExteriorCalculus.wedge_product — Function

Alias for the wedge product operator ∧.

CombinatorialSpaces.DiscreteExteriorCalculus.δ — Method

Codifferential operator from primal $n$ forms to primal $n-1$-forms.

CombinatorialSpaces.DiscreteExteriorCalculus.ℒ — Method

Lie derivative of $n$-form with respect to a vector field (or 1-form).

Specifically, this is the primal-dual Lie derivative defined in (Hirani 2003, Section 8.4) and (Desbrun et al 2005, Section 10).

CombinatorialSpaces.DiscreteExteriorCalculus.∇² — Method

Laplace-Beltrami operator on discrete forms.

This linear operator on primal $n$-forms defined by $∇² α := -δ d α$, where δ is the codifferential and d is the exterior derivative.

Note

For following texts such as Abraham-Marsden-Ratiu, we take the sign convention that makes the Laplace-Beltrami operator consistent with the Euclidean Laplace operator (the divergence of the gradient). Other authors, such as (Hirani 2003), take the opposite convention, which has the advantage of being consistent with the Laplace-de Rham operator Δ.

CombinatorialSpaces.DiscreteExteriorCalculus.♭ — Method

Flat operator converting vector fields to 1-forms.

A generic function for discrete flat operators. Currently only the DPP-flat from (Hirani 2003, Definition 5.5.2) and (Desbrun et al 2005, Definition 7.3) is implemented.

See also: the sharp operator ♯.

CombinatorialSpaces.DiscreteExteriorCalculus.♯ — Method

Sharp operator for converting 1-forms to vector fields.

A generic function for discrete sharp operators. Currently only the primal-primal flat from (Hirani 2003, Definition 5.8.1 and Remark 2.7.2) is implemented.

Note

A PP-flat is also defined in (Desbrun et al 2005, Definition 7.4) but differs in two ways: Desbrun et al's notation suggests a unit normal vector, whereas the gradient of Hirani's primal-primal interpolation function is not necessarily a unit vector. More importantly, Hirani's vector is a normal to a different face than Desbrun et al's, with further confusion created by the fact that Hirani's Figure 5.7 agrees with Desbrun et al's description rather than his own. That being said, to the best of our knowledge, our implementation is the correct one and agrees with Hirani's description, if not his figure.

See also: the flat operator ♭.