Simplicial sets

CombinatorialSpaces.SimplicialSets — Module

Simplicial sets in one, two, and three dimensions.

For the time being, this module provides data structures only for delta sets, also known as semi-simplicial sets. These include the face maps but not the degeneracy maps of a simplicial set. In the future we may add support for simplicial sets. The analogy to keep in mind is that graphs are to semi-simpicial sets as reflexive graphs are to simplicial sets.

Simplicial sets are inherently ordered structures. The "unordered" analogue of simplicial sets are symmetric simplicial sets, sometimes called just "symmetric sets." In one dimension, symmetric semi-simplicial sets are symmetric graphs.

This module does not implement symmetric simplicial sets as such. However, symmetric sets can be simulated with simplicial sets by enforcing that the ordering of the vertices of each face matches the ordering of the integer vertex IDs. The simplicial set then "presents" a symmetric set in a canonical way. The functions add_sorted_edge! and glue_sorted_triangle! automatically sort their inputs to ensure that the ordering condition is satisfied.

CombinatorialSpaces.SimplicialSets.AbstractDeltaSet1D — Type

Abstract type for 1D delta sets.

CombinatorialSpaces.SimplicialSets.AbstractDeltaSet2D — Type

Abstract type for 2D delta sets.

CombinatorialSpaces.SimplicialSets.DeltaSet1D — Type

A one-dimensional delta set, aka semi-simplicial set.

Delta sets in 1D are the same as graphs, and this type is just an alias for Graph. The face maps ∂ translates the graph-theoretic terminology into simplicial terminology.

CombinatorialSpaces.SimplicialSets.DeltaSet2D — Type

A 2D delta set, aka semi-simplicial set.

The triangles in a semi-simpicial set can be interpreted in several ways. Geometrically, they are triangles (2-simplices) whose three edges are directed according to a specific pattern, determined by the ordering of the vertices or equivalently by the simplicial identities. This geometric perspective is present through the subpart names ∂e0, ∂e1, and ∂e2 and through the boundary map ∂. Alternatively, the triangle can be interpreted as a higher-dimensional link or morphism, going from two edges in sequence (which might be called src2_first and src2_last) to a transitive edge (say tgt2). This is the shape of the binary composition operation in a category.

CombinatorialSpaces.SimplicialSets.E — Type

Edge in simplicial set: alias for Simplex{1}.

CombinatorialSpaces.SimplicialSets.EmbeddedDeltaSet1D — Type

A one-dimensional, embedded, oriented delta set.

CombinatorialSpaces.SimplicialSets.EmbeddedDeltaSet2D — Type

A two-dimensional, embedded, oriented delta set.

CombinatorialSpaces.SimplicialSets.OrientedDeltaSet1D — Type

A one-dimensional oriented delta set.

Edges are oriented from source to target when edge_orientation is true/positive and from target to source when it is false/negative.

CombinatorialSpaces.SimplicialSets.OrientedDeltaSet2D — Type

A two-dimensional oriented delta set.

Triangles are ordered in the cyclic order $(0,1,2)$ when tri_orientation is true/positive and in the reverse order when it is false/negative.

CombinatorialSpaces.SimplicialSets.Simplex — Type

Wrapper for simplex or simplices of dimension n.

See also: V, E, Tri.

CombinatorialSpaces.SimplicialSets.SimplexChain — Type

Wrapper for chain of oriented simplices of dimension n.

CombinatorialSpaces.SimplicialSets.SimplexForm — Type

Wrapper for discrete form, aka cochain, in simplicial set.

CombinatorialSpaces.SimplicialSets.Tri — Type

Triangle in simplicial set: alias for Simplex{2}.

CombinatorialSpaces.SimplicialSets.V — Type

Vertex in simplicial set: alias for Simplex{0}.

CombinatorialSpaces.SimplicialSets.add_sorted_edge! — Method

Add edge to simplicial set, respecting the order of the vertex IDs.

CombinatorialSpaces.SimplicialSets.add_sorted_edges! — Method

Add edges to simplicial set, respecting the order of the vertex IDs.

CombinatorialSpaces.SimplicialSets.add_triangle! — Method

Add a triangle (2-simplex) to a simplicial set, given its boundary edges.

In the arguments to this function, the boundary edges have the order $0 → 1$, $1 → 2$, $0 → 2$.

Warning

This low-level function does not check the simplicial identities. It is your responsibility to ensure they are satisfied. By contrast, triangles added using the function glue_triangle! always satisfy the simplicial identities, by construction. Thus it is often easier to use this function.

CombinatorialSpaces.SimplicialSets.boundary — Function

Alias for the face map and boundary operator ∂.

CombinatorialSpaces.SimplicialSets.coboundary — Function

Alias for the coboundary operator d.

CombinatorialSpaces.SimplicialSets.d — Method

The discrete exterior derivative, aka the coboundary operator.

CombinatorialSpaces.SimplicialSets.edge_vertices — Method

Boundary vertices of an edge.

CombinatorialSpaces.SimplicialSets.exterior_derivative — Function

Alias for the discrete exterior derivative d.

CombinatorialSpaces.SimplicialSets.glue_sorted_triangle! — Method

Glue a triangle onto a simplicial set, respecting the order of the vertices.

CombinatorialSpaces.SimplicialSets.glue_triangle! — Method

Glue a triangle onto a simplicial set, given its boundary vertices.

If a needed edge between two vertices exists, it is reused (hence the "gluing"); otherwise, it is created.

CombinatorialSpaces.SimplicialSets.nsimplices — Method

Number of simplices of given dimension in a simplicial set.

CombinatorialSpaces.SimplicialSets.orient! — Method

Consistently orient simplices in a simplicial set, if possible.

Two simplices with a common face are consistently oriented if they induce opposite orientations on the shared face. This function attempts to consistently orient all simplices of a given dimension and returns whether this has been achieved. Each connected component is oriently independently using the helper function orient_component!.

CombinatorialSpaces.SimplicialSets.orient_component! — Method

Consistently orient simplices in the same connected component, if possible.

Given an $n$-simplex and a choice of orientation for it, this function attempts to consistently orient all $n$-simplices that may be reached from it by traversing $(n-1)$-faces. The traversal is depth-first. If a consistent orientation is possible, the function returns true and the orientations are assigned; otherwise, it returns false and no orientations are changed.

If the simplicial set is not connected, the function orient! may be more convenient.

CombinatorialSpaces.SimplicialSets.orientation — Method

Orientation of simplex.

CombinatorialSpaces.SimplicialSets.point — Method

Point associated with vertex of complex.

CombinatorialSpaces.SimplicialSets.set_orientation! — Method

Set orientation of simplex.

CombinatorialSpaces.SimplicialSets.simplices — Method

Simplices of given dimension in a simplicial set.

CombinatorialSpaces.SimplicialSets.triangle_vertices — Method

Boundary vertices of a triangle.

This accessor assumes that the simplicial identities hold.

CombinatorialSpaces.SimplicialSets.volume — Method

$n$-dimensional volume of $n$-simplex spanned by given $n+1$ points.

CombinatorialSpaces.SimplicialSets.volume — Method

$n$-dimensional volume of $n$-simplex in an embedded simplicial set.

CombinatorialSpaces.SimplicialSets.∂ — Method

Face map and boundary operator on simplicial sets.

Given numbers n and 0 <= i <= n and a simplicial set of dimension at least n, the ith face map is implemented by the call

∂(n, i, s, ...)

The boundary operator on n-faces and n-chains is implemented by the call

∂(n, s, ...)

Note that the face map returns simplices, while the boundary operator returns chains (vectors in the free vector space spanned by oriented simplices).