Simplicial sets

CombinatorialSpaces.SimplicialSets — Module

Simplicial sets in one, two, and three dimensions.

For the time being, this module provides data structures only for semi-simplicial sets, which include the face maps but not the degeneracy maps. In the future we may add support for simplicial sets having both face and degeneracy maps. The analogy to keep in mind is that graphs are to semi-simpicial sets as reflexive graphs are to simplicial sets.

Simplicial sets are inherently ordered structures. The "unordered" version of simplicial sets are symmetric simplicial sets, sometimes called just "symmetric sets." In one dimension, symmetric semi-simplicial sets are symmetric graphs.

This module does not directly implement symmetric simplicial sets. However, they can be simulated with simplicial sets by enforcing that the ordering of the vertices of each face matches the ordering of the integer vertex IDs. The simplicial set then "presents" a symmetric simplicial set in a canonical way. The functions add_sorted_edge! and glue_sorted_triangle! automatically sort their inputs to ensure that the ordering condition is satisfied.

CombinatorialSpaces.SimplicialSets.AbstractSemiSimplicialSet1D — Type

Abstract type for 1D semi-simplicial sets.

CombinatorialSpaces.SimplicialSets.AbstractSemiSimplicialSet2D — Type

Abstract type for 2D semi-simplicial sets.

CombinatorialSpaces.SimplicialSets.OrientedSimplicialSet1D — Type

A one-dimensional oriented simplicial set.

Edges are oriented from source to target when edge_orientation is true/positive and from target to source when it is false/negative.

CombinatorialSpaces.SimplicialSets.OrientedSimplicialSet2D — Type

A two-dimensional oriented simplicial set.

Triangles are ordered in the cyclic order $(0,1,2)$ (with numbers defined by triangle_vertex) when tri_orientation is true/positive and in the reverse order when it is false/negative.

CombinatorialSpaces.SimplicialSets.SemiSimplicialSet1D — Type

A one-dimensional semi-simplicial set.

Semi-simplicial sets in 1D are the same as graphs, and this type is just an alias for Graph. The boundary operator ∂₁ translates the graph-theoretic terminology into simplicial terminology.

CombinatorialSpaces.SimplicialSets.SemiSimplicialSet2D — Type

A 2D semi-simplicial set.

The triangles in a simpicial set can be interpreted in several ways. Geometrically, they are triangles (2-simplices) whose three edges are directed according to a specific pattern, determined by the ordering of the vertices or equivalently by the simplicial identities. This geometric perspective is encoded by the boundary map [∂₂](@ref). Alternatively, the triangle can be seen as a higher-dimensional link or morphism, going from two edges in sequence (here called src2_first and src2_last) to a transitive edge (tgt2). Not coincidentally, this is the shape of the binary composition operation in a category. The categorical interpretation is the one used in the definition of the schema.

CombinatorialSpaces.SimplicialSets.add_sorted_edge! — Method

Add edge to simplicial set, respecting the order of the vertex IDs.

CombinatorialSpaces.SimplicialSets.add_sorted_edges! — Method

Add edges to simplicial set, respecting the order of the vertex IDs.

CombinatorialSpaces.SimplicialSets.add_triangle! — Method

Add a triangle (2-simplex) to a simplicial set, given its boundary edges.

Warning

This low-level function does not check the simplicial identities. It is your responsibility to ensure they are satisfied. By contrast, triangles added using the function glue_triangle! always satisfy the simplicial identities, by the nature of the construction.

CombinatorialSpaces.SimplicialSets.edge_sign — Method

Sign (±1) associated with edge orientation.

CombinatorialSpaces.SimplicialSets.glue_sorted_triangle! — Method

Glue a triangle onto a simplicial set, respecting the order of the vertices.

CombinatorialSpaces.SimplicialSets.glue_triangle! — Method

Glue a triangle onto a simplicial set, given its boundary vertices.

If a needed edge between two vertices exists, it is reused (hence the "gluing"); otherwise, it is created.

CombinatorialSpaces.SimplicialSets.triangle_sign — Method

Sign (±1) associated with triangle orientation.

CombinatorialSpaces.SimplicialSets.triangle_vertex — Method

Boundary vertex of a triangle.

This accessor assumes that the simplicial identities hold.

CombinatorialSpaces.SimplicialSets.∂ — Method

Face map and boundary operator on simplicial sets.

Given numbers n and 0 <= i <= n and a simplicial set of dimension at least n, the ith face map is implemented by the call

∂(n, i, s, ...)

The boundary operator on n-faces and n-chains is implemented by the call

∂(n, s, ...)

Note that the face map returns simplices, while the boundary operator returns chains (vectors in the free vector space spanned by oriented simplices).

CombinatorialSpaces.SimplicialSets.∂₁ — Method

Boundary operator on edges or 1-chains in simplicial set.

CombinatorialSpaces.SimplicialSets.∂₂ — Method

Boundary operator on triangles or 2-chains in simplicial set.